Updated Builtin Shaders to 2018.1.0f2

+++ CGIncludes/UnityStandardInput.cginc
@@ -38,8 +38,8 @@ half        _DetailNormalMapScale;
 sampler2D   _SpecGlossMap;
 sampler2D   _MetallicGlossMap;
 half        _Metallic;
-half        _Glossiness;
-half        _GlossMapScale;
+float       _Glossiness;
+float       _GlossMapScale;
 
 sampler2D   _OcclusionMap;
 half        _OcclusionStrength;
